refine collective strategy
